Amend code Auth and CRUD post

+++ cmd/api/auth.go
@@ -20,10 +20,21 @@ type RegisterUserPayload struct {
 }
 
 type UserWithToken struct {
-	*store.User
+	Username string`json:"username" validate:"required,max=100"`
+	Email	string`json:"email" validate:"required,email,max=255"`
+	*store.Role
 	Token string `json:"token"`
 }
 
+
+type RegisterUserData struct {
+	ID        uint   
+	Username string `json:"username" validate:"required,max=100"`
+	Email    string `json:"email" validate:"required,email,max=255"`
+	Password string `json:"password" validate:"required,min=3,max=72"`
+	IsActive  bool    
+	RoleID    uint
+}
 
 // registerUserHandler godoc
 //
@@ -36,28 +47,25 @@ type UserWithToken struct {
 //	@Success		201		{object}	UserWithToken		"User registered"
 //	@Failure		400		{object}	error
 //	@Failure		500		{object}	error
-//	@Router			/authentication/user [post]
+//	@Router			/auth/user [post]
 func (app *application) registerUserHandler(c *fiber.Ctx) error {
 	var payload RegisterUserPayload
+	ctx := c.Context()
 
-	// Parse JSON
 	if err := c.BodyParser(&payload); err != nil {
 		return app.badRequestResponse(c, err)
 	}
-
+	
+	role, err := app.store.Roles.GetByName(ctx, "user")
+	if err != nil {
+		return app.internalServerError(c, err)
+	}
 	user := &store.User{
 		Username: payload.Username,
 		Email:    payload.Email,
-		Role: store.Role{
-			Name: "user",
-		},
+		IsActive: false,
+		RoleID: role.ID,
 	}
-
-	if err := user.Password.Set(payload.Password); err != nil {
-		return app.internalServerError(c, err)
-	}
-
-	ctx := c.Context()
 
 	plainToken := uuid.New().String()
 	hash := sha256.Sum256([]byte(plainToken))
@@ -65,34 +73,69 @@ func (app *application) registerUserHandler(c *fiber.Ctx) error {
 
 
 	userWithToken := UserWithToken{
-		User:  user,
+		Username:  user.Username,
+		Email: user.Email,
+		Role: role,
 		Token: hashToken,
 	}
 
 	activationURL := fmt.Sprintf("%s/confirm/%s", app.config.frontendURL, plainToken)
-
 	isProdEnv := app.config.env == "production"
 
-	mailVars := struct {
-		Username      string
-		ActivationURL string
-	}{
-		Username:      user.Username,
-		ActivationURL: activationURL,
-	}
-
-	status, err := app.mailer.Send(mailer.UserWelcomeTemplate, user.Username, user.Email, mailVars, !isProdEnv)
+	err = app.store.Users.Create(ctx, user, payload.Password)
 	if err != nil {
-		app.logger.Errorw("error sending welcome email", "error", err)
-
-		if err := app.store.Users.Delete(ctx, user.ID); err != nil {
-			app.logger.Errorw("error deleting user", "error", err)
+		switch err {
+		case store.ErrDuplicateEmail:
+			app.badRequestResponse(c, err)
+		case store.ErrDuplicateUsername:
+			app.badRequestResponse(c, err)
+		default:
+			app.internalServerError(c, err)
+		}
+		return err
+	}
+	if  app.config.mail.enabled{ 
+		mailVars := struct {
+			Username      string
+			ActivationURL string
+		}{
+			Username:      user.Username,
+			ActivationURL: activationURL,
 		}
 
-		return app.internalServerError(c, err)
-	}
+		status, err := app.mailer.Send(mailer.UserWelcomeTemplate, user.Username, user.Email, mailVars, !isProdEnv)
+		if err != nil {
+			app.logger.Errorw("error sending welcome email", "error", err)
 
-	app.logger.Infow("Email sent", "status code", status)
+			if err := app.store.Users.Delete(ctx, user.ID); err != nil {
+				app.logger.Errorw("error deleting user", "error", err)
+			}
+
+			return app.internalServerError(c, err)
+		}
+
+		app.logger.Infow("Email sent", "status code", status)
+	}else{
+		gtUer, err := app.store.Users.GetByEmail(ctx, user.Email)
+		if err != nil {
+		switch err {
+		case store.ErrNotFound:
+			return app.notFoundResponse(c, err)
+		default:
+			return app.internalServerError(c, err)
+		}
+	}
+		if err := app.store.Users.Activate(ctx, gtUer.ID); err != nil { 
+			switch err {
+		case store.ErrNotFound:
+			return app.notFoundResponse(c, err)
+		default:
+			app.logger.Errorw("error updating user status after mailer skip", "error", err, "userID", user.ID)
+			return app.internalServerError(c, err)
+			}
+		}
+	}
+	
 
 	return c.Status(fiber.StatusCreated).JSON(userWithToken)
 }
@@ -114,16 +157,14 @@ type CreateUserTokenPayload struct {
 //	@Failure		400		{object}	error
 //	@Failure		401		{object}	error
 //	@Failure		500		{object}	error
-//	@Router			/authentication/token [post]
+//	@Router			/auth/token [post]
 func (app *application) createTokenHandler(c *fiber.Ctx) error {
 	var payload CreateUserTokenPayload
 
-	// Parse request
 	if err := c.BodyParser(&payload); err != nil {
 		return app.badRequestResponse(c, err)
 	}
 
-	// Fetch user
 	user, err := app.store.Users.GetByEmail(c.Context(), payload.Email)
 	if err != nil {
 		switch err {
@@ -133,15 +174,13 @@ func (app *application) createTokenHandler(c *fiber.Ctx) error {
 			return app.internalServerError(c, err)
 		}
 	}
-
-	// Compare password
-	if err := user.Password.Compare(payload.Password); err != nil {
+	if err := user.Authenticate(payload.Password); err != nil {
 		return app.unauthorizedErrorResponse(c, err)
 	}
 
-	// JWT claims
 	claims := jwt.MapClaims{
 		"sub": user.ID,
+		"role": user.Role.Name,
 		"exp": time.Now().Add(app.config.auth.token.exp).Unix(),
 		"iat": time.Now().Unix(),
 		"nbf": time.Now().Unix(),
